elasticsearch-dsl-example/requirements.txt
elasticsearch-dsl==7.1.0